More work on the settings page
We will need more stuff in the database for users: an alt email and 3
notification settings (what are they?). We also need a table that stores
available times for users. The way I planned this in terms of database
schema is to have a new table, "Times", with columns like this:

|userID|year|month|day|hour|

userID is a foreign key. year, month, day and hour are NOT NULL and must
be set. The four of them represent a time in which the user represented by
userID is avaiable. These four date columns can simply be ints.